Fix errors and cli parsing in optimize_videos

+++ programs/optimize_videos.sh
@@ -21,14 +21,21 @@
 #
 # ⚠ Warning: this tool overwrites the source videos! Backup your videos before using this program.
 
-NAME="optimize_videos"
+NAME="optimize_videos.sh"
 ERROR_RESIZE="Incorrect value for --resize. Turning resize off."
 
 scale=""
 no_audio=0
 
-# SOURCING FILES
-. ./lib/utils.sh
+# Debug tools
+FORMAT_NORMAL=$(tput sgr0)
+FORMAT_BOLD=$(tput bold)
+
+is_dry_run=1
+
+format_bold() {
+	printf "%s%s%s" "$FORMAT_BOLD" "$*" "$FORMAT_NORMAL"
+}
 
 echo_help() {
 	printf 'Compress videos to h264 MP4 with ffmpeg.
@@ -45,19 +52,22 @@ No positional arguments.
 -r/--resize size  -- Resize the video to this size using the scale filter. The size must be of a form
  supported by the ffmpeg scale filter. For example, 1280:720 for a 720p size or iw/2:-1 to divide
  the original video size by two. See https://trac.ffmpeg.org/wiki/Scaling for more information.
--n/--no-audio     -- Remove all audio from the output videos.'
-	"$(format_bold Usage)" "$(format_bold Positional arguments)" "$(format_bold Options)"
+-n/--no-audio     -- Remove all audio from the output videos.' "$(format_bold "Usage")" "$(format_bold "Positional arguments")" "$(format_bold "Options")"
 	exit 0
 }
 
 parse_cli_arguments() {
-	arguments=$(getopt --name "$NAME" -o "h,r,n:" -l "help,resize:,no-audio" -- "$@")
+	arguments=$(getopt --name "$NAME" -o "h,d,r:,n" -l "help,dry-run,resize:,no-audio" -- "$@")
 	eval set -- "$arguments"
 	while true; do
 		case "$1" in
 		-h | --help)
 			echo_help
 			break
+			;;
+		-d | --dry-run)
+			is_dry_run=1
+			shift
 			;;
 		-r | --resize)
 			echo "$2" | grep -E ".+:.+" && scale="$2" || echo "$ERROR_RESIZE"
@@ -71,31 +81,39 @@ parse_cli_arguments() {
 			shift
 			break
 			;;
+		*)
+			echo "There was an error"
+			exit 1
+			;;
 		esac
+	done
+	for i in "$@"; do
+		echo "$i"
 	done
 }
 
 # Compresses and overwrites
 # Arguments:
 # $@: list of paths to video files
-compress_video() {
+compress_videos() {
 	command="ffmpeg -hwaccel auto -y -v quiet -i %s -c:v libx264 -crf 20 -preset slow"
 	test $no_audio -eq 1 && command="$command -an" || command="$command -c:a aac -b:a 320k"
 	test scale != "" && command="$command -filter \"scale=$scale\""
 
-	for filepath in "$@"; do
+	while read -r filepath; do
 		filename=$(basename "$filepath")
-		path_temp="$(dirname "$filepath")/temp_$(echo "$filename" | sed 's/\.[A-Za-z0-9]+$/.mp4/')"
-		path_out="$(echo $path_temp | sed 's/temp_//')"
+		path_temp=$(dirname "$filepath")/temp_$(echo "$filename" | sed 's/\.[A-Za-z0-9]+$/.mp4/')
+		path_out=$(echo "$path_temp" | sed 's/temp_//')
 		ffmpeg_command=$(printf "$command %s" "$filepath" "$path_temp")
-		command "$ffmpeg_command"
-		mv -v "$path_temp" "$path_out"
-	done
+		test $is_dry_run -eq 0 && command "$ffmpeg_command" || echo "$ffmpeg_command"
+		test $is_dry_run -eq 0 && mv -v "$path_temp" "$path_out" || echo Moving "$path_temp" to "$path_out"
+	done <"$1"
 }
 
 main() {
-	parse_cli_arguments "$@"
-	test $? -eq 0 && compress_lossy "$@"
+	temp_file=$(mktemp)
+	parse_cli_arguments "$@" >"$temp_file"
+	test $? -eq 0 && compress_videos "$temp_file"
 	exit $?
 }
